GUI/Login
I have created a Jlabel that can print that can print an error message if the user forget to enter a username or password. I have created this method by making get methods for the login and passwordfield and a set method for the login failure message in the the login class.

LOGIC/Snakeapp
In the Snakeapp i have created a private boolean value. The boolean value is a value with two choices. I used this method to check if the password and player id field is filled out or empty.
After that i am using the actionlistener i have created for login to creating a if else statement. I am getting the player field and password field to check if they are filled out if not i am printing the Jlabel i created in GUI/Login with the failure message.

// Prabdeep